PiaBlogv2/backend/requirements.txt
2to3==1.0
aioredis==1.2.0
alabaster==0.7.11
anaconda-client==1.7.2
anaconda-project==0.8.2
appdirs==1.4.3
asgiref==2.3.2
asn1crypto==0.24.0
astroid==2.1.0
astropy==3.0.4
async-timeout==3.0.1
atomicwrites==1.2.1
attrs==18.2.0
autobahn==19.3.2
Automat==0.7.0
autopep8==1.4.4
Babel==2.6.0
backcall==0.1.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.3
bitarray==0.8.3
bkcharts==0.2
blaze==0.11.3
bleach==2.1.4
bokeh==0.13.0
boto==2.49.0
Bottleneck==1.2.1
certifi==2018.11.29
cffi==1.11.5
channels==2.1.7
channels-redis==2.3.3
chardet==3.0.4
click==6.7
cloudpickle==0.5.5
clyent==1.2.2
colorama==0.4.1
comtypes==1.1.7
constantly==15.1.0
contextlib2==0.5.5
cryptography==2.4.2
cycler==0.10.0
Cython==0.28.5
cytoolz==0.9.0.1
daphne==2.2.5
dask==0.19.1
datashape==0.5.4
decorator==4.3.0
defusedxml==0.5.0
distributed==1.23.1
Django==2.1.5
django-allauth==0.38.0
django-contrib-comments==1.9.0
django-cors-headers==2.5.0
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-filer==1.3.2
django-filter==2.0.0
django-mptt==0.8.7
django-polymorphic==2.0.3
django-rest-auth==0.9.3
django-tinymce4-lite==1.7.4
django-tinymce4-widget==4.0.0
djangorestframework==3.9.0
docutils==0.14
easy-thumbnails==2.5
entrypoints==0.2.3
et-xmlfile==1.0.1
fastcache==1.0.2
filebrowser-safe==0.5.0
filelock==3.0.8
Flask==1.0.2
Flask-Cors==3.0.6
future==0.17.1
gevent==1.3.6
glob2==0.6
grappelli-safe==0.5.1
greenlet==0.4.15
gspread==3.1.0
gunicorn==19.9.0
h5py==2.8.0
heapdict==1.0.0
hiredis==1.0.0
html5lib==1.0.1
httplib2==0.12.0
hyperlink==18.0.0
idna==2.7
imageio==2.4.1
imagesize==1.1.0
incremental==17.5.0
ipykernel==4.9.0
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.1
isort==4.3.4
itsdangerous==0.24
jdcal==1.4
jedi==0.12.1
Jinja2==2.10
jsmin==2.2.2
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyterlab==0.34.9
jupyterlab-launcher==0.13.1
keyring==13.2.1
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
llvmlite==0.24.0
locket==0.2.0
lxml==4.2.5
Markdown==3.0.1
MarkupSafe==1.0
matplotlib==2.2.3
mccabe==0.6.1
menuinst==1.4.14
Mezzanine==4.3.1
mezzanine-theme-moderna==0.2
mistune==0.8.3
mkl-fft==1.0.4
mkl-random==1.0.1
more-itertools==4.3.0
mpmath==1.0.0
msgpack==0.6.1
multipledispatch==0.6.0
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.1
nltk==3.3
nose==1.3.7
notebook==5.6.0
numba==0.39.0
numexpr==2.6.8
numpy==1.15.1
numpydoc==0.8.0
oauth2client==4.1.2
oauthlib==3.0.1
odo==0.5.1
olefile==0.46
openpyxl==2.5.6
packaging==17.1
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
partd==0.3.8
path.py==11.1.0
pathlib2==2.3.2
patsy==0.5.0
pep8==1.7.1
pickleshare==0.7.4
Pillow==5.4.1
pkginfo==1.4.2
pluggy==0.7.1
ply==3.11
prometheus-client==0.3.1
prompt-toolkit==1.0.15
psutil==5.4.7
psycopg2==2.7.7
py==1.6.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0.2
pyflakes==2.0.0
pygal==2.4.0
Pygments==2.2.0
PyHamcrest==1.9.0
pylint==2.2.2
pyodbc==4.0.24
pyOpenSSL==18.0.0
pyparsing==2.2.0
PySocks==1.6.8
pytest==3.8.0
pytest-arraydiff==0.2
pytest-astropy==0.4.0
pytest-doctestplus==0.1.3
pytest-openfiles==0.3.0
pytest-remotedata==0.3.0
python-dateutil==2.7.3
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.9
PyWavelets==1.0.0
pywin32==223
pywinpty==0.5.4
PyYAML==3.13
pyzmq==17.1.2
QtAwesome==0.4.4
qtconsole==4.4.1
QtPy==1.5.0
requests==2.20.1
requests-oauthlib==1.2.0
rope==0.11.0
rsa==3.4.2
ruamel-yaml==0.15.46
scikit-image==0.14.0
scikit-learn==0.19.2
scipy==1.1.0
seaborn==0.9.0
selenium==3.141.0
Send2Trash==1.5.0
service-identity==17.0.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.12.0
snowballstemmer==1.2.1
sortedcollections==1.0.1
sortedcontainers==2.0.5
Sphinx==1.7.9
sphinxcontrib-websupport==1.1.0
spyder==3.3.1
spyder-kernels==0.2.6
SQLAlchemy==1.2.11
sqlparse==0.2.4
statsmodels==0.9.0
stripe==2.15.0
sympy==1.2
tables==3.4.4
tblib==1.3.2
terminado==0.8.1
testpath==0.3.1
toolz==0.9.0
tornado==5.1
tqdm==4.26.0
traitlets==4.3.2
Twisted==18.7.0
txaio==18.8.1
typed-ast==1.1.1
tzlocal==1.5.1
unicodecsv==0.14.1
Unidecode==0.4.21
urllib3==1.24.1
virtualenv==16.4.3
virtualenv-clone==0.5.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.1
win-inet-pton==1.0.1
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.11.0
xlrd==1.1.0
XlsxWriter==1.1.0
xlwings==0.11.8
xlwt==1.3.0
yapf==0.27.0
zict==0.1.3
zope.interface==4.5.0